requirements.txt
flask>=3.0
openai>=1.0.0
pdfplumber>=0.11
python-dotenv>=1.0
requests>=2.31
